fix(test-helper): Fix Sinon/QUnit Assertion Issue
This change wraps the pass/fail from sinon.assert in a Qunit.ok assertion. This makes it
so that Qunit detects sinon.assert as an actual assertion.

sinon.assert methods are not detected as QUnit Assertions. When no qunit assertions exist
in the test, Qunit will throw an error:

"Expected at least one assertion, but none were run - call expect(0) to accept zero assertions."

This is expected behavior from Qunit, but when we assert with sinon.assert, there is technically
assertion there.